filetool.sh: Alert the user when backup fails
Backup operation currently fails silently--the only difference between successful backup and failed backup is that successful one outputs "Done." at end of operation. Because of the silent failures, I've had a non-existent files in my .filetool.lst for several months now. With this commit, filetool.sh now alerts the user that there was a problem, and tells user where to look for clues.